fixup! pkgimages: relocation-lean image format and replay-driven load path
Review fixes: close the WaitEntryN testset the rebase left open; restore the
uint64_t checksum parameter so the sysimage build-id marker survives; read
`owner` through the accessor in the two chain walks that tested it before
materializing; root the foreign intersection vector during activation replay;
mark entries the prepass bails out on as dirty; and require an atom, not just
a concrete type, before declaring a call-signature slot disjoint from a
foreign method (Tuple{DataType} is concrete but intersects Tuple{Type{Int}}).

+++ src/staticdata.c
@@ -5770,7 +5770,7 @@ static int all_usings_unchanged_implicit(jl_module_t *mod)
 }
 
 static void jl_restore_system_image_from_stream_(ios_t *f, jl_image_t *image,
-                                                 jl_array_t *depmods, uint32_t checksum,
+                                                 jl_array_t *depmods, uint64_t checksum,
                                 /* outputs */    jl_array_t **restored JL_REQUIRE_ROOTED_SLOT,
                                                  jl_array_t **init_order JL_REQUIRE_ROOTED_SLOT,
                                                  jl_array_t **extext_methods JL_REQUIRE_ROOTED_SLOT,
